ft: rename `GenerationOptions` to `ShortIdOptions` and relocate it to project root
